Split off OpenLoco.h
