+++ src/OpenLoco/Audio/Audio.cpp
@@ -12,7 +12,7 @@
 #include "../Objects/ObjectManager.h"
 #include "../Objects/SoundObject.h"
 #include "../Objects/TreeObject.h"
-#include "../OpenLoco.h"
+#include "../SceneManager.h"
 #include "../Ui/WindowManager.h"
 #include "../Utility/Stream.hpp"
 #include "../Vehicles/Vehicle.h"
