+++ src/OpenLoco/GameCommands/LoadSaveQuit.cpp
@@ -1,6 +1,6 @@
 #include "../Audio/Audio.h"
 #include "../Game.h"
-#include "../OpenLoco.h"
+#include "../SceneManager.h"
 #include "../Tutorial.h"
 #include "../Ui/WindowManager.h"
 #include "../Ui/WindowType.h"
