+++ src/OpenLoco/GameCommands/RemoveBuilding.cpp
@@ -5,8 +5,8 @@
 #include "../Localisation/StringIds.h"
 #include "../Map/TileManager.h"
 #include "../Objects/BuildingObject.h"
-#include "../OpenLoco.h"
 #include "../S5/S5.h"
+#include "../SceneManager.h"
 #include "../TownManager.h"
 #include "GameCommands.h"
