+++ src/OpenLoco/GameCommands/SetGameSpeed.cpp
@@ -1,5 +1,5 @@
 #include "../Interop/Interop.hpp"
-#include "../OpenLoco.h"
+#include "../SceneManager.h"
 #include "GameCommands.h"
 
 using namespace OpenLoco::Interop;
