+++ src/OpenLoco/Input/Keyboard.cpp
@@ -8,7 +8,7 @@
 #include "../Interop/Interop.hpp"
 #include "../Intro.h"
 #include "../Localisation/StringIds.h"
-#include "../OpenLoco.h"
+#include "../SceneManager.h"
 #include "../Tutorial.h"
 #include "../Ui.h"
 #include "../Ui/Screenshot.h"
