+++ src/OpenLoco/Input/MouseInput.cpp
@@ -10,7 +10,7 @@
 #include "../Map/TileManager.h"
 #include "../Objects/ObjectManager.h"
 #include "../Objects/RoadObject.h"
-#include "../OpenLoco.h"
+#include "../SceneManager.h"
 #include "../StationManager.h"
 #include "../TownManager.h"
 #include "../Tutorial.h"
