+++ src/OpenLoco/Input/ShortcutManager.cpp
@@ -5,8 +5,8 @@
 #include "../Interop/Interop.hpp"
 #include "../LastGameOptionManager.h"
 #include "../Localisation/StringIds.h"
-#include "../OpenLoco.h"
 #include "../S5/S5.h"
+#include "../SceneManager.h"
 #include "../StationManager.h"
 #include "../TownManager.h"
 #include "../Ui/WindowManager.h"
