+++ src/OpenLoco/Network/Network.cpp
@@ -4,8 +4,8 @@
 #include "../GameCommands/GameCommands.h"
 #include "../GameState.h"
 #include "../Graphics/Gfx.h"
-#include "../OpenLoco.h"
 #include "../ScenarioManager.h"
+#include "../SceneManager.h"
 #include "NetworkClient.h"
 #include "NetworkServer.h"
 #include "Socket.h"
