+++ src/OpenLoco/Network/NetworkClient.cpp
@@ -2,9 +2,9 @@
 #include "../Config.h"
 #include "../Console.h"
 #include "../GameCommands/GameCommands.h"
-#include "../OpenLoco.h"
 #include "../Platform/Platform.h"
 #include "../S5/S5.h"
+#include "../SceneManager.h"
 #include "../Ui/WindowManager.h"
 #include "../Utility/Stream.hpp"
 #include "NetworkConnection.h"
