+++ src/OpenLoco/Network/NetworkServer.cpp
@@ -3,10 +3,10 @@
 #include "../Core/Span.hpp"
 #include "../GameCommands/GameCommands.h"
 #include "../GameState.h"
-#include "../OpenLoco.h"
 #include "../Platform/Platform.h"
 #include "../S5/S5.h"
 #include "../ScenarioManager.h"
+#include "../SceneManager.h"
 #include "../Utility/Stream.hpp"
 #include "../Utility/String.hpp"
 #include "NetworkConnection.h"
