+++ src/OpenLoco/StationManager.cpp
@@ -8,8 +8,8 @@
 #include "Localisation/StringIds.h"
 #include "Map/TileManager.h"
 #include "Objects/IndustryObject.h"
-#include "OpenLoco.h"
 #include "ScenarioManager.h"
+#include "SceneManager.h"
 #include "TownManager.h"
 #include "Ui/WindowManager.h"
 #include "Window.h"
