+++ src/OpenLoco/Ui/ViewportInteraction.cpp
@@ -16,8 +16,8 @@
 #include "../Objects/TrackObject.h"
 #include "../Objects/TreeObject.h"
 #include "../Objects/WallObject.h"
-#include "../OpenLoco.h"
 #include "../Paint/Paint.h"
+#include "../SceneManager.h"
 #include "../StationManager.h"
 #include "../TownManager.h"
 #include "../Ui.h"
