+++ src/OpenLoco/Ui/WindowManager.cpp
@@ -13,7 +13,7 @@
 #include "../Map/Tile.h"
 #include "../Map/TileManager.h"
 #include "../MultiPlayer.h"
-#include "../OpenLoco.h"
+#include "../SceneManager.h"
 #include "../StationManager.h"
 #include "../TownManager.h"
 #include "../Tutorial.h"
