+++ src/OpenLoco/Vehicles/VehicleHead.cpp
@@ -26,6 +26,7 @@
 #include "../Objects/VehicleObject.h"
 #include "../OpenLoco.h"
 #include "../ScenarioManager.h"
+#include "../SceneManager.h"
 #include "../StationManager.h"
 #include "../TownManager.h"
 #include "../Ui/WindowManager.h"
