+++ src/OpenLoco/Windows/CompanyWindow.cpp
@@ -17,9 +17,9 @@
 #include "../Objects/CompetitorObject.h"
 #include "../Objects/InterfaceSkinObject.h"
 #include "../Objects/ObjectManager.h"
-#include "../OpenLoco.h"
 #include "../Scenario.h"
 #include "../ScenarioObjective.h"
+#include "../SceneManager.h"
 #include "../Ui/Dropdown.h"
 #include "../Ui/ScrollView.h"
 #include "../Ui/WindowManager.h"
