+++ src/OpenLoco/Windows/Construction/OverheadTab.cpp
@@ -10,7 +10,7 @@
 #include "../../Objects/RoadObject.h"
 #include "../../Objects/TrackExtraObject.h"
 #include "../../Objects/TrackObject.h"
-#include "../../OpenLoco.h"
+#include "../../SceneManager.h"
 #include "../../Ui/Dropdown.h"
 #include "../../Widget.h"
 #include "Construction.h"
