+++ src/OpenLoco/Windows/MapWindow.cpp
@@ -17,7 +17,7 @@
 #include "../Objects/ObjectManager.h"
 #include "../Objects/RoadObject.h"
 #include "../Objects/TrackObject.h"
-#include "../OpenLoco.h"
+#include "../SceneManager.h"
 #include "../StationManager.h"
 #include "../TownManager.h"
 #include "../Types.hpp"
