+++ src/OpenLoco/Windows/News/Common.cpp
@@ -7,7 +7,7 @@
 #include "../../MessageManager.h"
 #include "../../Objects/InterfaceSkinObject.h"
 #include "../../Objects/ObjectManager.h"
-#include "../../OpenLoco.h"
+#include "../../SceneManager.h"
 #include "News.h"
 
 using namespace OpenLoco::Interop;
