+++ src/OpenLoco/Windows/Options.cpp
@@ -12,8 +12,8 @@
 #include "../Objects/CurrencyObject.h"
 #include "../Objects/InterfaceSkinObject.h"
 #include "../Objects/ObjectManager.h"
-#include "../OpenLoco.h"
 #include "../S5/S5.h"
+#include "../SceneManager.h"
 #include "../Ui.h"
 #include "../Ui/Dropdown.h"
 #include "../Ui/WindowManager.h"
