+++ src/OpenLoco/Windows/ProgressBar.cpp
@@ -4,7 +4,7 @@
 #include "../Graphics/ImageIds.h"
 #include "../Interop/Interop.hpp"
 #include "../Localisation/StringIds.h"
-#include "../OpenLoco.h"
+#include "../SceneManager.h"
 #include "../Ui/WindowManager.h"
 #include "../Widget.h"
 #include "../Window.h"
