+++ src/OpenLoco/Windows/ScenarioSelect.cpp
@@ -8,9 +8,9 @@
 #include "../Localisation/FormatArguments.hpp"
 #include "../Localisation/StringIds.h"
 #include "../Objects/ObjectManager.h"
-#include "../OpenLoco.h"
 #include "../Scenario.h"
 #include "../ScenarioManager.h"
+#include "../SceneManager.h"
 #include "../Ui/ScrollView.h"
 #include "../Ui/WindowManager.h"
 #include "../Widget.h"
