+++ src/OpenLoco/Windows/TerraForm.cpp
@@ -20,8 +20,8 @@
 #include "../Objects/TreeObject.h"
 #include "../Objects/WallObject.h"
 #include "../Objects/WaterObject.h"
-#include "../OpenLoco.h"
 #include "../Scenario.h"
+#include "../SceneManager.h"
 #include "../Ui/Dropdown.h"
 #include "../Ui/ScrollView.h"
 #include "../Ui/WindowManager.h"
