+++ src/OpenLoco/Windows/TimePanel.cpp
@@ -12,8 +12,8 @@
 #include "../Network/Network.h"
 #include "../Objects/InterfaceSkinObject.h"
 #include "../Objects/ObjectManager.h"
-#include "../OpenLoco.h"
 #include "../ScenarioObjective.h"
+#include "../SceneManager.h"
 #include "../Ui.h"
 #include "../Ui/Dropdown.h"
 #include "../Ui/WindowManager.h"
