+++ src/OpenLoco/Windows/ToolbarTop.cpp
@@ -19,8 +19,8 @@
 #include "../Objects/RoadObject.h"
 #include "../Objects/TrackObject.h"
 #include "../Objects/WaterObject.h"
-#include "../OpenLoco.h"
 #include "../S5/S5.h"
+#include "../SceneManager.h"
 #include "../StationManager.h"
 #include "../TownManager.h"
 #include "../Ui/Dropdown.h"
